+++ packages/interaction/src/Resizer.svelte
@@ -1,12 +1,20 @@
 <script>
     import {getContext} from 'svelte';
+    import {bgEvent, helperEvent} from '@event-calendar/common';
 
     export let event;
 
-    let {theme, _resizable} = getContext('state');
+    let {theme, eventDurationEditable, editable} = getContext('state');
+
+    let resizable;
+    $: resizable = !bgEvent(event.display) &&
+        !helperEvent(event.display) && (
+            (event.durationEditable ?? $eventDurationEditable) ||
+            (event.editable ?? $editable)
+        )
 </script>
 
-{#if event.display === 'auto' && $_resizable(event)}
+{#if resizable}
     <div
         class="{$theme.resizer}"
         on:pointerdown